Show top communities when logged out, misc fixes
