+++ src/helpers/array.ts
@@ -1,11 +1,5 @@
 import { pullAt } from "lodash";
 
-export function notEmpty<TValue>(
-  value: TValue | null | undefined,
-): value is TValue {
-  return value !== null && value !== undefined;
-}
-
 // https://stackoverflow.com/a/60132060/1319878
 export const arrayOfAll =
   <T>() =>
